Relationship doesn’t need to be polymorphic.
I clung onto this idea for a while, but actually it just makes things
complicated. Particularly the API REST framework, which doesn’t seem to
play well with django-polymorphic.

We don’t really need to get a mixed bag of relationships for a
particular actor. It seems cool, but we don’t use it on the frontend
currently anyway.